# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.createWFHPeriodRequest:
    properties:
      end_date:
        type: string
      reason:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - start_date
    type: object
...
  models.Role:
    enum:
    - employee
    - team_lead
    - admin
    - logistics
    type: string
    x-enum-varnames:
    - RoleEmployee
    - RoleTeamLead
    - RoleAdmin
    - RoleLogistics
  services.CreateUserInput:
    properties:
      default_meal_preference:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        $ref: '#/definitions/models.Role'
    required:
    - email
    - name
    - password
    - role
    type: object
  services.UpdateUserInput:
    properties:
      default_meal_preference:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        $ref: '#/definitions/models.Role'
    type: object
host: localhost:8080
info:
  contact:
    email: support@craftsbite.com
    name: API Support
  description: CraftsBite Backend API for meal management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CraftsBite API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get list of all users (Admin/Logistics only)
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account (Admin only)
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error or creation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Deactivate a user account (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deactivated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Deactivation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate user
      tags:
      - users
    get:
      description: Get user details (Admin or own profile)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user details (Admin or own profile)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error or update failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
...
  /users/me/team:
    get:
      description: Get team details for authenticated employee
      produces:
      - application/json
      responses:
        "200":
          description: Team retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - only employees
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my team
      tags:
      - users
  /users/me/team-members:
    get:
      description: Get team members for authenticated team lead
      produces:
      - application/json
      responses:
        "200":
          description: Team members retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - only team leads
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my team members
      tags:
      - users
  /wfh-periods:
    get:
      description: Get all work-from-home periods (Admin/Logistics only)
      produces:
      - application/json
      responses:
        "200":
          description: WFH periods retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List WFH periods
      tags:
      - wfh-periods
    post:
      consumes:
      - application/json
      description: Create a new work-from-home period (Admin/Logistics only)
      parameters:
      - description: WFH period details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.createWFHPeriodRequest'
      produces:
      - application/json
      responses:
        "201":
          description: WFH period created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create WFH period
      tags:
      - wfh-periods
  /wfh-periods/{id}:
    delete:
      description: Delete a work-from-home period (Admin/Logistics only)
      parameters:
      - description: WFH Period ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: WFH period deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete WFH period
      tags:
      - wfh-periods
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"